+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+# 4.1.7
+
+* Automatically extend EBS volume when size changes
+* Add AWS Backup support to EBS volume
+* Add CloudWatch alarm support to root and data volumes
+
 # 4.1.6
 
 * Change EBS volume type to gp3